refactor static pages to server components
